Move weight and feature values from dict to array.
Instead of keeping every features weight in the dictionary, now a index
is kept instead, pointing into where the actual weight is stored in a
numpy array. This way we won't have to do list comprehensions to create
an array every time before computing output.

To make this work, feature output values had to be stored in the same
way, with the same index. Now, instead of creating a new array each time
we calculate the feature values, we just update the global one.